mode set for returning user
