+++ src/App.js
@@ -1,16 +1,90 @@
 import "./App.css";
-import Nav from "./components/Nav";
+// import Nav from "./components/_Nav";
 import Header from "./components/Header";
 import Main from "./components/Main";
 import Footer from "./components/Footer";
 
+import { useState, useEffect } from "react";
+
+
+// /*Light Mode and Dark Mode Toggle
+//   -create a function to get local storage theme value and set it to a variable
+const getStorageMode = () => {
+  //   -let the default theme be light
+  let mode = 'lightMode';
+  if (localStorage.getItem('mode')) {
+    mode = localStorage.getItem('mode');
+  }
+  return mode;
+};
+
 
 function App() {
+  
+  //   -create state variable to store the theme
+  const [mode, setMode] = useState(getStorageMode());
+  
+  //   -create a function to toggle the theme
+  const toggleMode = () => {
+    //     -if the theme is dark, set the localStorage theme to dark
+    if (mode === 'lightMode') {
+      setMode('darkMode');
+      //     -else the theme is light, set the localStorage theme to light
+    } else {
+      setMode('lightMode');
+    }
+  };
+  
+  //     -use the useEffect hook to set the theme
+  useEffect(() => {
+    document.documentElement.className = mode;
+    localStorage.setItem('mode', mode);
+  }, [mode]);
+  
+  //initial state is false
+  const [showAbout, setShowAbout] = useState(false);
+  
+  //toggle the state
+  const toggleAbout = () => {
+    setShowAbout(!showAbout);
+  };
+
   return (
     <div className="App">
-      <Nav />
+      <nav>
+        <div className="navContainer">
+          <ul>
+            <li>
+              <button onClick={toggleAbout}>
+                {
+                  showAbout 
+                  ? "X" 
+                  : "Help"
+                } 
+              </button>
+              <p>
+                {
+                  showAbout
+                  ? "Type in your thought and click the button to share them with the world."
+                  : ""
+                }
+              </p>            
+            </li>
+            <li>
+            {/* //   -create a toggle button to toggle the theme on click */}
+              <button className="toggleMode" onClick={toggleMode}>
+                {
+                  //write ternary operator to toggle the button text
+                  mode === 'lightMode' ? 'Dark Mode' : 'Light Mode'
+                }
+              </button>
+            </li>
+          </ul>
+        </div>
+      </nav>
+
       <Header />
-      <Main />
+      <Main mode={mode} />
       <Footer />
     </div>
   );
