+++ src/partials/_header.scss
@@ -13,6 +13,7 @@ header {
   h1 {
     padding: 1rem;
     margin: 1rem auto;
+    font-family: $brandFont;
   }
   .imgContainer {
     width: 50%;
