+++ src/partials/_variables.scss
@@ -10,3 +10,4 @@ $black: rgb(34, 33, 33);
 
 //fonts
 $font: 'Georgia', serif;
+$brandFont: Comic Sans MS, Comic Sans, cursive;
